Reformatted changes, removed unecessary files.
